feat: 포인트 적립 메서드 추가
- pointBalance Long -> int 수정 : 사용 포인트와 타입 일치 시키기 위해 수정함
- 수정으로 인한 리팩토링

+++ src/test/java/dev/nbcsparta/assignment/nbccheckoutsystem/order/service/OrderCreateTest.java
@@ -314,7 +314,7 @@ class OrderCreateTest {
     private Members member(Long id, Long pointBalance) {
         Members member = new Members("test@test.com", "password", "name", "010-0000-0000");
         ReflectionTestUtils.setField(member, "id", id);
-        ReflectionTestUtils.setField(member, "pointBalance", pointBalance);
+        ReflectionTestUtils.setField(member, "pointBalance", pointBalance.intValue());
         return member;
     }
